repack : 리팩토링 문서 정리
